+++ NEWS.md
@@ -5,6 +5,9 @@
   
 * `x$name` never attempts to evaluate `name` (#1368).
 
+* Namespaced dplyr calls now error if the function doesn't exist, or 
+  a translation is not available (#1426).
+
 * `db_sql_render()` correctly passes on `...` when re-calling with 
   `sql_options` set (#1394).